[okd-core] Default SERVER_FLAGS.registry
As kubevirt-web-ui handles deployment of the V2V as a temporary workaround
for missing other more suitable "setup component" to do that, the code needs
to be know where/what container images for V2V use.

Future work: either move the deployment logic out of web-ui
or configure registry via ConfigMap and bridge process.

The later option is alligned with changes in this commit.